add javafx ui

+++ src/service/CarService.java
@@ -1,12 +1,15 @@
 package service;
 
 import decorator.PromotionDecorator;
+import exceptions.CarDeletionException;
 import exceptions.CarNotFoundException;
 import exceptions.InvalidPromotionException;
 import model.Car;
 import model.Promotion;
 import repository.CarRepository;
 import repository.PromotionRepository;
+import repository.SaleRepository;
+import repository.ServiceAppointmentRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -14,14 +17,23 @@ import java.util.stream.Collectors;
 public class CarService {
     private final CarRepository carRepository;
     private final PromotionRepository promotionRepository;
+    private final SaleRepository saleRepository;
+    private final ServiceAppointmentRepository serviceAppointmentRepository;
 
     public CarService() {
-        this(new CarRepository(), new PromotionRepository());
+        this(new CarRepository(), new PromotionRepository(), new SaleRepository(), new ServiceAppointmentRepository());
     }
 
     public CarService(CarRepository carRepository, PromotionRepository promotionRepository) {
+        this(carRepository, promotionRepository, new SaleRepository(), new ServiceAppointmentRepository());
+    }
+
+    public CarService(CarRepository carRepository, PromotionRepository promotionRepository,
+                      SaleRepository saleRepository, ServiceAppointmentRepository serviceAppointmentRepository) {
         this.carRepository = carRepository;
         this.promotionRepository = promotionRepository;
+        this.saleRepository = saleRepository;
+        this.serviceAppointmentRepository = serviceAppointmentRepository;
     }
 
     public Car addCar(Car car) {
@@ -30,8 +42,14 @@ public class CarService {
         return saved;
     }
 
-    public void removeCar(int carId) throws CarNotFoundException {
+    public void removeCar(int carId) throws CarNotFoundException, CarDeletionException {
         findCarById(carId);
+        if (saleRepository.hasSalesForCar(carId)) {
+            throw new CarDeletionException(
+                    "Mașina nu poate fi ștearsă deoarece are vânzări înregistrate.");
+        }
+        promotionRepository.deleteByCarId(carId);
+        serviceAppointmentRepository.deleteByCarId(carId);
         carRepository.deleteById(carId);
         AuditService.getInstance().logAction("REMOVE_CAR");
     }
@@ -41,6 +59,13 @@ public class CarService {
         car.setPrice(newPrice);
         carRepository.update(car);
         AuditService.getInstance().logAction("UPDATE_PRICE");
+    }
+
+    public Car updateCar(Car car) throws CarNotFoundException {
+        findCarById(car.getId());
+        Car updated = carRepository.update(car);
+        AuditService.getInstance().logAction("UPDATE_CAR");
+        return updated;
     }
 
     public List<Car> searchCars(String brand, Double minPrice, Double maxPrice, Integer year) {
